backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
python-multipart
openai-whisper
reportlab
torch
httpx
